introduce NamedInfo as super class of AnnotatedInfo
